requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.3
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
kombu==4.1.0
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.2
Pillow==4.2.1
psycopg2==2.7.3
PyJWT==1.5.2
pylibmc==1.5.2
pytz==2017.2
qrcode==5.3
redis==2.10.5
requests==2.18.2
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4

